Re-enable data manipulator unit tests. Fix all manipulator unit tests.

+++ src/main/java/org/spongepowered/common/data/manipulator/mutable/entity/SpongeFoodData.java
@@ -36,10 +36,10 @@ import org.spongepowered.common.data.value.SpongeValueFactory;
 public class SpongeFoodData extends AbstractData<FoodData, ImmutableFoodData> implements FoodData {
 
     private int foodLevel;
-    private float foodSaturationLevel;
-    private float foodExhaustionLevel;
+    private double foodSaturationLevel;
+    private double foodExhaustionLevel;
 
-    public SpongeFoodData(int foodLevel, float foodSaturationLevel, float foodExhaustionLevel) {
+    public SpongeFoodData(int foodLevel, double foodSaturationLevel, double foodExhaustionLevel) {
         super(FoodData.class);
         this.foodLevel = foodLevel;
         this.foodSaturationLevel = foodSaturationLevel;
